+++ test/SystemTextJson.JsonDiffPatch.UnitTests/NodeTests/DiffTests.cs
@@ -49,5 +49,113 @@ namespace SystemTextJson.JsonDiffPatch.UnitTests.NodeTests
 
             Assert.Equal("{\"_t\":\"a\",\"_0\":[\"\",1,3]}", diff!.ToJsonString());
         }
+
+        [Fact]
+        public void Diff_NullProperty()
+        {
+            var left = JsonNode.Parse("{\"a\":1}");
+            var right = JsonNode.Parse("{\"a\":null}");
+
+            var diff = left.Diff(right);
+
+            Assert.Equal("{\"a\":[1,null]}", diff!.ToJsonString());
+        }
+
+        [Fact]
+        public void Diff_NullArrayItem()
+        {
+            var left = JsonNode.Parse("[1]");
+            var right = JsonNode.Parse("[null]");
+
+            var diff = left.Diff(right);
+
+            Assert.Equal("{\"_t\":\"a\",\"_0\":[1,0,0],\"0\":[null]}", diff!.ToJsonString());
+        }
+
+        [Fact]
+        public void PropertyFilter_LeftProperty()
+        {
+            var left = JsonNode.Parse("{\"a\":1}");
+            var right = JsonNode.Parse("{}");
+
+            var diff = left.Diff(right, new JsonDiffOptions
+            {
+                PropertyFilter = (prop, _) => !string.Equals(prop, "a")
+            });
+
+            Assert.Null(diff);
+        }
+
+        [Fact]
+        public void PropertyFilter_RightProperty()
+        {
+            var left = JsonNode.Parse("{}");
+            var right = JsonNode.Parse("{\"a\":1}");
+
+            var diff = left.Diff(right, new JsonDiffOptions
+            {
+                PropertyFilter = (prop, _) => !string.Equals(prop, "a")
+            });
+
+            Assert.Null(diff);
+        }
+
+        [Fact]
+        public void PropertyFilter_NestedProperty()
+        {
+            var left = JsonNode.Parse("{\"foo\":{\"bar\":{\"a\":1}}}");
+            var right = JsonNode.Parse("{\"foo\":{\"bar\":{\"a\":2}}}");
+
+            var diff = left.Diff(right, new JsonDiffOptions
+            {
+                PropertyFilter = (prop, _) => !string.Equals(prop, "a")
+            });
+
+            Assert.Null(diff);
+        }
+
+        [Fact]
+        public void PropertyFilter_ArrayItem()
+        {
+            var left = JsonNode.Parse("[{\"a\":1}]");
+            var right = JsonNode.Parse("[{\"a\":2}]");
+
+            var diff = left.Diff(right, new JsonDiffOptions
+            {
+                PropertyFilter = (prop, _) => !string.Equals(prop, "a")
+            });
+
+            Assert.Null(diff);
+        }
+
+        [Fact]
+        public void PropertyFilter_IgnoreByPath()
+        {
+            var left = JsonNode.Parse("{\"a\":{\"b\":{\"c\":1}}}");
+            var right = JsonNode.Parse("{\"a\":{\"b\":{\"c\":2}}}");
+
+            var diff = left.Diff(right, new JsonDiffOptions
+            {
+                PropertyFilter = (_, ctx) =>
+                    ctx.Left<JsonNode>().GetPath() != "$.a.b" && ctx.Right<JsonNode>().GetPath() != "$.a.b"
+            });
+
+            Assert.Null(diff);
+        }
+
+        [Fact]
+        public void PropertyFilter_ArrayItem_ExplicitFallbackMatch()
+        {
+            var left = JsonNode.Parse("[{\"a\":1}]");
+            var right = JsonNode.Parse("[{\"a\":2}]");
+
+            var diff = left.Diff(right, new JsonDiffOptions
+            {
+                ArrayObjectItemMatchByPosition = true,
+                PropertyFilter = (prop, _) => !string.Equals(prop, "a")
+            });
+
+            Assert.Null(diff);
+        }
     }
 }